Further organization of arguments in CLI help text
