+++ Sources/MarkersExtractorCLI/MarkersExtractorCLI+General.swift
@@ -24,38 +24,6 @@ extension MarkersExtractorCLI {
         var enableSubframes: Bool = MarkersExtractor.Settings.Defaults.enableSubframes
         
         @Option(
-            help: ArgumentHelp(
-                "Annotations to import. If captions are used, their start timecode determines their position.",
-                valueName: caseIterableValueString(for: MarkersSource.self)
-            )
-        )
-        var markersSource: MarkersSource = MarkersExtractor.Settings.Defaults.markersSource
-        
-        @Flag(
-            help: ArgumentHelp(
-                "For chapter markers, use their thumbnail pin position for thumbnail image generation."
-            )
-        )
-        var useChapterMarkerThumbnails: Bool = MarkersExtractor.Settings.Defaults.useChapterMarkerThumbnails
-        
-        @Option(
-            name: [.customLong("exclude-role")],
-            help: ArgumentHelp(
-                "Exclude markers with a specified role. This argument can be supplied more than once to apply multiple role exclusions.",
-                valueName: "name"
-            )
-        )
-        var excludeRoles: [String] = []
-        
-        @Flag(
-            name: [.customLong("include-disabled")],
-            help: ArgumentHelp(
-                "Include markers on disabled clips. By default, disabled clips are ignored."
-            )
-        )
-        var includeDisabled: Bool = MarkersExtractor.Settings.Defaults.includeDisabled
-        
-        @Option(
             name: [.customLong("folder-format")],
             help: ArgumentHelp(
                 "Output folder name format.",
